+++ samples/src/main/kotlin/jp/shiguredo/sora/sample/ui/ScreencastSetupActivity.kt
@@ -50,6 +50,7 @@ class ScreencastSetupActivity : AppCompatActivity() {
     }
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+        super.onActivityResult(requestCode, resultCode, data)
         screencastStarter?.onActivityResult(requestCode, resultCode, data)
     }