+++ samples/src/main/kotlin/jp/shiguredo/sora/sample/ui/VideoChatRoomActivity.kt
@@ -2,6 +2,8 @@ package jp.shiguredo.sora.sample.ui
 
 import android.annotation.TargetApi
 import android.content.Context
+import android.content.pm.ActivityInfo
+import android.content.res.Configuration
 import android.content.res.Resources
 import android.graphics.Color
 import android.media.AudioManager
@@ -23,6 +25,7 @@ import jp.shiguredo.sora.sdk.channel.data.ChannelAttendeesCount
 import jp.shiguredo.sora.sdk.channel.option.SoraAudioOption
 import jp.shiguredo.sora.sdk.channel.option.SoraVideoOption
 import jp.shiguredo.sora.sdk.error.SoraErrorReason
+import jp.shiguredo.sora.sdk.util.SoraLogger
 import kotlinx.android.synthetic.main.activity_video_chat_room.*
 import org.webrtc.PeerConnection
 import org.webrtc.SurfaceViewRenderer
@@ -41,6 +44,7 @@ class VideoChatRoomActivity : AppCompatActivity() {
     private var audioCodec:  SoraAudioOption.Codec = SoraAudioOption.Codec.OPUS
     private var audioEnabled = true
     private var audioBitRate: Int? = null
+    private var audioStereo: Boolean = false
     private var videoBitRate: Int? = null
     private var videoWidth: Int = SoraVideoOption.FrameSize.Portrait.VGA.x
     private var videoHeight: Int = SoraVideoOption.FrameSize.Portrait.VGA.y
@@ -48,13 +52,17 @@ class VideoChatRoomActivity : AppCompatActivity() {
     private var fps: Int = 30
     private var fixedResolution = false
     private var clientId: String? = null
-    private var sdpSemantics = PeerConnection.SdpSemantics.UNIFIED_PLAN
 
     private var oldAudioMode: Int = AudioManager.MODE_INVALID
 
     private var streamType = SoraStreamType.BIDIRECTIONAL
 
     private var ui: VideoChatRoomActivityUI? = null
+
+    override fun onConfigurationChanged(newConfig: Configuration) {
+        SoraLogger.d(TAG, "onConfigurationChanged: orientation=${newConfig?.orientation}")
+        super.onConfigurationChanged(newConfig)
+    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         Log.d(TAG, "onCreate")
@@ -123,14 +131,20 @@ class VideoChatRoomActivity : AppCompatActivity() {
             else       -> false
         }
 
-        videoBitRate = when (intent.getStringExtra("VIDEO_BIT_RATE")) {
+        videoBitRate = when (val stringValue = intent.getStringExtra("VIDEO_BIT_RATE")) {
             "UNDEFINED" -> null
-            else -> intent.getStringExtra("VIDEO_BIT_RATE")?.toInt()
+            else -> stringValue?.toInt()
         }
 
-        audioBitRate = when (intent.getStringExtra("AUDIO_BIT_RATE")) {
+        audioBitRate = when (val stringValue = intent.getStringExtra("AUDIO_BIT_RATE")) {
             "UNDEFINED" -> null
-            else -> intent.getStringExtra("AUDIO_BIT_RATE")?.toInt()
+            else -> stringValue?.toInt()
+        }
+
+        audioStereo = when (intent.getStringExtra("AUDIO_STEREO")) {
+            "MONO"   -> false
+            "STEREO" -> true
+            else     -> false
         }
 
         clientId = when (intent.getStringExtra("CLIENT_ID")) {
@@ -140,10 +154,17 @@ class VideoChatRoomActivity : AppCompatActivity() {
             "RANDOM UUID" -> UUID.randomUUID().toString()
             else -> null
         }
-        sdpSemantics = when (intent.getStringExtra("SDP_SEMANTICS")) {
-            "Unified Plan" -> PeerConnection.SdpSemantics.UNIFIED_PLAN
-            "Plan B"       -> PeerConnection.SdpSemantics.PLAN_B
-            else           -> PeerConnection.SdpSemantics.UNIFIED_PLAN
+
+        // ステレオでは landscape にしたほうが内蔵マイクを使うときに自然な向きとなる。
+        // それ以外は、リモート映像の分割が簡単になるように portrait で動かす。
+        if (audioStereo) {
+            if(resources.configuration.orientation == Configuration.ORIENTATION_PORTRAIT) {
+                requestedOrientation = ActivityInfo.SCREEN_ORIENTATION_LANDSCAPE
+            }
+        } else {
+            if(resources.configuration.orientation == Configuration.ORIENTATION_LANDSCAPE) {
+                requestedOrientation = ActivityInfo.SCREEN_ORIENTATION_PORTRAIT
+            }
         }
 
         ui = VideoChatRoomActivityUI(
@@ -256,7 +277,7 @@ class VideoChatRoomActivity : AppCompatActivity() {
                 audioEnabled      = audioEnabled,
                 audioCodec        = audioCodec,
                 audioBitRate      = audioBitRate,
-                sdpSemantics      = sdpSemantics,
+                audioStereo       = audioStereo,
                 streamType        = streamType,
                 clientId          = clientId,
                 listener          = channelListener,